Revamp repo and use github.com/shirou/gopsutil
Apologies for the monolithic commit, but it's hard to make atomic
commits with such a big change.

Here's why the changes happened: If you run the `cmd/main.go` (which was
in `osversion/main.go`before the revamp) on a go 1.18 compiler, you'll
get this error:

```
cgo: malformed DWARF TagVariable entry
make[1]: *** [app] Error 2
```

I dug and found that just including `#include <stdlib.h>` is enough to
trigger the error with CGO. No idea why this is a problem but it is. It
could be because we should do `//go:build darwin && cgo`, not just `//go:build darwin`.

Anyways, I found that less code is always better and shirou's gopsutil
library does 95% of what we need, except fetching the Android API / OS
Version bit, which is mainly the only addition this library does. At one
point, I'll make the changes here upstream to shirou's repo and use
everything from there.

Main changes to this repo:

- Add a go.mod
- Remove all unused functions in our codebase
  - We just use `GetHumanReadable()`, not `GetString()` or `GetSemanticVersion`
- Remove all platform-specific functions in lieu of gopsutil
  - Except the Android bits
- Add a working `./cmd` for all platforms
  - The one before it didn't work on Android
- Remove the regex test
  - The test was just checking if `GetString()` was called on a Darwin
    system, it'll have `darwin` in the string. That job is now delegated
    to gopsutil, which has an extensive CI test that runs on many
    platforms. Ours doesn't have a CI test and it's quite redundant to
    retest the same thing.